refactor(merkle-tree): split mmcs.rs into a feature-organized mmcs/ module
`mmcs.rs` had grown to 3175 lines bundling four concerns behind one file.
Split it into an `mmcs/` folder organized by role, following the std layout
(a type's inherent impl in one place; trait impls and free functions in their
own files):

- mod.rs       MerkleTreeMmcs struct + its full inherent impl
               (new, cap_height, replay_arity_and_positions,
                proof_arity_schedule, open_batch_pruned, verify_batch_pruned)
- batch.rs     impl Mmcs for MerkleTreeMmcs (commit / open / verify)
- error.rs     MerkleTreeError, PrunedProofError
- geometry.rs  check_widths, validate_commit_reachable_heights, ArityAndPositions
- pruned.rs    PrunedBatchOpening

Pure mechanical refactor: no behavior, API, or visibility change. The public
surface (lib.rs re-exports) and the `crate::mmcs::check_widths` use in
hiding_mmcs.rs resolve unchanged; `validate_commit_reachable_heights` is
widened from private to `pub(crate)` only so two sibling modules can call it.
Every relocated function body is byte-for-byte identical to the original, and
unit tests are co-located by feature with each module defining its own
fixtures (no shared test-only types file).

All 102 tests pass; fmt, clippy, and doc are clean.
